ref: create abstract directory initialization service in plugin api
