+++ plugins/widoco/src/main/java/cz/lukaskabc/ontology/ontopus/plugin/widoco/service/init/FilesDirectoryInitializationService.java
@@ -1,43 +1,15 @@
 package cz.lukaskabc.ontology.ontopus.plugin.widoco.service.init;
 
-import cz.lukaskabc.ontology.ontopus.api.service.core.InitializationService;
-import cz.lukaskabc.ontology.ontopus.core_model.exception.InitializationException;
+import cz.lukaskabc.ontology.ontopus.api.service.init.DirectoryInitializationService;
 import cz.lukaskabc.ontology.ontopus.plugin.widoco.config.WidocoPluginConfig;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
-import java.io.File;
-import java.nio.file.Files;
-
 /** Ensures that {@link WidocoPluginConfig#filesDirectory} exists and is writable and readable directory */
-@Order(500)
+@Order(DirectoryInitializationService.ORDER)
 @Component
-public class FilesDirectoryInitializationService implements InitializationService {
-    private final WidocoPluginConfig config;
-
+public class FilesDirectoryInitializationService extends DirectoryInitializationService {
     public FilesDirectoryInitializationService(WidocoPluginConfig config) {
-        this.config = config;
-    }
-
-    @Override
-    public void initialize() {
-        final File filesDirectory = config.getFilesDirectory().toFile();
-        if (!filesDirectory.exists()) {
-            final boolean success = filesDirectory.mkdirs();
-            if (!success) {
-                throw new InitializationException(
-                        "Failed to create files directory at " + filesDirectory.getAbsolutePath());
-            }
-        }
-        if (!filesDirectory.isDirectory()) {
-            throw new InitializationException(
-                    "Files directory is missing or not a directory: " + filesDirectory.getAbsolutePath());
-        }
-        if (!Files.isWritable(filesDirectory.toPath())) {
-            throw new InitializationException("Files directory is not writable: " + filesDirectory.getAbsolutePath());
-        }
-        if (!Files.isReadable(filesDirectory.toPath())) {
-            throw new InitializationException("Files directory is not readable: " + filesDirectory.getAbsolutePath());
-        }
+        super(config.getFilesDirectory());
     }
 }
